+++ test/SlashHandler.t.sol
@@ -394,7 +394,7 @@ contract TestableSlashHandler is SlashHandler, Redeemer {
     __readStub__();
   }
 
-  function convertToReceiptTokenAmount(uint256, /* reservePoolId_ */ uint256 /*reserveAssetAmount_ */ )
+  function convertToReceiptTokenAmount(uint8, /* reservePoolId_ */ uint256 /*reserveAssetAmount_ */ )
     public
     view
     override
@@ -403,7 +403,7 @@ contract TestableSlashHandler is SlashHandler, Redeemer {
     __readStub__();
   }
 
-  function convertToReserveAssetAmount(uint256, /* depositReceiptTokenAmount_ */ uint256 /* reservePoolId_ */ )
+  function convertToReserveAssetAmount(uint8, /* reservePoolId_ */ uint256 /* depositReceiptTokenAmount_ */ )
     public
     view
     override
